fixup! runtime: add hypervisor options for NUMA topology
